requirements.txt
alembic==1.3.2
aniso8601==8.0.0
apturl==0.5.2
asn1crypto==0.24.0
astroid==2.3.3
attrs==19.3.0
autopep8==1.4.4
backcall==0.1.0
bcrypt==3.1.7
beautifulsoup4==4.7.1
blinker==1.4
Brlapi==0.6.7
cached-property==1.5.1
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
Click==7.0
colorama==0.3.7
command-not-found==0.3
cryptography==2.3
cupshelpers==1.0
decorator==4.4.1
defer==1.0.6
distro==1.3.0
distro-info===0.21ubuntu2
docformatter==1.3.1
docker==3.4.1
docker-compose==1.21.0
docker-pycreds==0.3.0
dockerpty==0.4.1
docopt==0.6.2
entrypoints==0.3
environ==1.0
environs==7.1.0
flake8==3.7.9
Flask==1.1.1
Flask-Admin==1.5.4
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.8
Flask-JWT==0.3.2
Flask-JWT-Extended==3.24.1
Flask-Login==0.4.1
Flask-Migrate==2.5.2
Flask-MySQLdb==0.2.0
Flask-OAuthlib==0.9.5
Flask-RESTful==0.3.7
flask-restplus==0.13.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.1
Flask-Testing==0.7.1
html5lib==1.0.1
httplib2==0.11.3
idna==2.8
importlib-metadata==1.3.0
ipython==5.8.0
ipython-genutils==0.2.0
isort==4.3.21
itsdangerous==1.1.0
jedi==0.15.2
Jinja2==2.10.3
jsonschema==3.2.0
jwcrypto==0.6.0
keyring==17.1.1
keyrings.alt==3.1.1
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lazy-object-proxy==1.4.3
louis==3.8.0
lxml==4.3.3
macaroonbakery==1.2.1
Mako==1.1.0
MarkupSafe==1.1.1
marshmallow==3.3.0
mccabe==0.6.1
more-itertools==8.0.2
mysqlclient==1.4.6
netifaces==0.10.4
oauthlib==2.1.0
olefile==0.46
parso==0.5.2
pexpect==4.7.0
pickleshare==0.7.5
Pillow==5.4.1
prompt-toolkit==1.0.15
protobuf==3.6.1
ptyprocess==0.6.0
pycairo==1.16.2
pycodestyle==2.5.0
pycparser==2.19
pycrypto==2.6.1
pycups==1.9.73
pyflakes==2.1.1
Pygments==2.5.2
PyGObject==3.32.0
pyinotify==0.9.6
PyJWT==1.7.1
pylint==2.4.4
pymacaroons==0.13.0
PyMySQL==0.9.3
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyRFC3339==1.1
pyrsistent==0.15.6
python-apt===1.8.5-ubuntu0.3
python-dateutil==2.8.1
python-debian==0.1.34
python-dotenv==0.10.3
python-editor==1.0.4
python-jwt==3.2.5
pytz==2019.3
pyxdg==0.25
PyYAML==3.13
reportlab==3.5.18
requests==2.22.0
requests-oauthlib==1.3.0
requests-unixsocket==0.1.5
SecretStorage==2.3.1
simplegeneric==0.8.1
simplejson==3.16.0
six==1.13.0
soupsieve==1.8
SQLAlchemy==1.3.12
system-service==0.3
systemd-python==234
texttable==1.6.0
traitlets==4.3.3
twilio==6.33.1
typed-ast==1.4.0
ubuntu-advantage-tools==19.2
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
untokenize==0.1.1
urllib3==1.25.7
usb-creator==0.3.3
virtualenv==16.7.9
wadllib==1.3.3
wcwidth==0.1.7
webencodings==0.5.1
websocket-client==0.53.0
Werkzeug==0.16.0
wrapt==1.11.2
WTForms==2.2.1
xkit==0.0.0
zipp==0.6.0
